% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittedRanef.R
\name{FittedRanefScalar}
\alias{FittedRanefScalar}
\title{Construct a Fitted random effect object for a scalar parameter of interest}
\usage{
FittedRanefScalar(fixef, fixef_se, sd, fdr_q, primerid = NA_character_,
  converge = NA, method, walltime, coretime, obs_used = NA)
}
\arguments{
\item{fixef}{vector of fixed effects, eg one per gene}

\item{fixef_se}{standard errors for fixed effects}

\item{sd}{vector of population variance components}

\item{fdr_q}{FDR q-values}

\item{method}{}

\item{walltime}{}

\item{coretime}{}

\item{obs_used}{}
}
\value{
data.table
}
\description{
Construct a Fitted random effect object for a scalar parameter of interest
}
