% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{simulateOne}
\alias{simulateOne}
\title{Simulate from a single feature/gene}
\usage{
simulateOne(X, beta, random_dm, ranef_likelihood, phi_fun = phi_constant(),
  obs_likelihood = ObsLikelihood(normal_likelihood))
}
\arguments{
\item{X}{fixed effect design matrix}

\item{beta}{fixed effect coefficients}

\item{random_dm}{a RandomDesignMatrix object}

\item{ranef_likelihood}{random effect generator function, object of class RanefLikelihood}

\item{phi_fun}{dispersion generator function}

\item{obs_likelihood}{the observation generator, object of class ObsLikelihood}
}
\value{
vector of observations Y, linear predictor eta, random effects b
}
\description{
Simulate from a single feature/gene
}
